requirements.txt
Django>=4.0.2
djangorestframework>=3.13.1
mysqlclient>=2.1.0
redis>=4.2.0
social-auth-app-django>=5.0.0
lockfile>=0.12.2
decouple>=0.0.7
